+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [1.11.13] - unreleased
 ### Added
+- Optional `token_cmd` to the reclone.yaml configuration
 ### Changed
 ### Deprecated
 ### Removed